backend/requirements_new.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
firebase-admin==6.2.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
numpy==1.24.3
python-dotenv==1.0.0
twilio==8.10.0
fpdf==1.7.2
langchain==0.1.0
langchain-google-genai==0.0.6
requests==2.31.0
Pillow==10.1.0
pytesseract==0.3.10
